fix no size() value of the vector returned by binary file reader
Former-commit-id: 511f2abaf797f07c8960b5794924ea5028b40b63 [formerly 3feb13063ba99ae2a0dbb158d05c9d8836c50557]
Former-commit-id: 64329f47161059962defc479671713c50551cb6e
Former-commit-id: 032f914a6af3d9d870f929bad1a6d0ba40e52898 [formerly bf0cef9fab4bcca6c86da3a464a667d140155a0a] [formerly 4a10004a13f5cc79c1f2d2226fe8e72c8a328ad8 [formerly 49d8845ef5a6e1126f7eed9074c7e012c84f3022]] [formerly c04ba9b7a2e517e441ecb25489baa79e26d2c40b [formerly 8083065c83aa61e688ee416bc2c22741727620bc] [formerly 1da59871fc6471e196ba5d570af1568193d51f11 [formerly 202f51826c101ceb75978f8802b643cbf4f6675c]]]
Former-commit-id: 65bf0b1bb67d591271e53f9b8854522444f22b15 [formerly 10a2ce5045f22b85e620ee4123c6bf6101dc09cf] [formerly b71979c4b54de4c771b31b9364f4b676cc3809b2 [formerly 4cd1114c9e0fb1020e00faaad0b57b9e59c6b4d6]]
Former-commit-id: 3a378fd71e3b43e005250db9bf377350ec6d4758 [formerly aa2ea32a0637a074362677a611bde0e99929520f]
Former-commit-id: 891321279d91702fb77f33ba0ee90f2aafe35b3a

+++ src/utility.hpp
@@ -38,7 +38,7 @@ std::vector<Request> LoadRequestsFromCsvFile(std::string path_to_csv);
 std::vector<Pos> LoadNetworkNodesFromCsvFile(std::string path_to_csv);
 
 ///// \brief A function loading the precomputed minimum mean travel time path of each node pair from a csv file.
-std::vector<std::vector<int>> LoadShortestPathTableFromCsvFile(std::string path_to_csv);
+std::vector<std::vector<size_t>> LoadShortestPathTableFromCsvFile(std::string path_to_csv);
 //std::vector<std::vector<int>> LoadShortestPathTableFromCsvFile(
 //        std::string path_to_nodes_csv, std::string path_to_path_table_csv);
